requirements.txt
# 舟山定海渔港雷达监控系统 V2.0
# 依赖列表

# ========================
# 核心依赖（必须安装）
# ========================

# 串口通信
pyserial>=3.5

# NMEA协议解析
pynmea2>=1.19.0

# Web框架
flask>=2.3.0
flask-socketio>=5.3.0

# 数据处理
numpy>=1.24.0

# ========================
# 可选依赖
# ========================

# 性能监控
psutil>=5.9.0

# 开发测试
pytest>=7.4.0
pytest-cov>=4.1.0

# 代码检查（可选）
bandit>=1.7.0
